OK I have cleaned a bunch of stuff up to make the code more readible. I've also changed the --no2d option to be --yes2d - and we default with no 2D surfacs. Edit hdf2.c to add your own 2d fields if you want them.
The most significant changes involved parsedir.c where I created new
subroutines in place of redundant blocks of code. It reads much easier
now.

If no --x0 --x1 --y0 --y1 arguments are provided, it now defaults to the
actual (x0,y0) -> (x1,y1) that was saved, rather than assuming x0=y0=0
etc. (I rarely save the full domain). It was a PITA to have to go in the
directories and find the smallest x0, largest x1 etc. Now it's done as
part of the metadata retrieval and is cached in the .cm1hdf5_blah files
(which I should change to contain the basename of the cm1hdf5 files in
case you are in a directory where there is more than one dataset being
operated on... that's a TODO...

Another TODO is to get z1 metadata - need to save it in CM1 first. Right
now it's just the size of the Z dimension of the saved 3D arrays....

I'm thinking about writing a routine that just goes out and assesses the
LOFS metadata (like the stuff that prints out in hdf2nc and makevisit).
Somethign like makevisit, actually, but more user friendly and sent to
the screen. I believe I will write it in python. This low-level C stuff
runs fast but requires lots of coding. Stay tuned.

 Changes to be committed:
	modified:   hdf2.c
	modified:   parsedir.c