I spent a lot of time squinting at the screen confused but I think I have it straightened out. First, I looked more closely at uh and uf and tweaked their calculations (these should really just be written to the /mesh group in CM1 - that's a TODO). Mallocs should also match what is used and I think I got rid of buffer overruns. I changed the macro for these as I do not believe they need the +1 offset, looking at how they are used in CM1.
As far as the 3D buffers go, I now have adoped to having all of the
macros for BUF, TEM, and TEM1 (and the pointers) all look the same
regarding array indexing and array size. This is something that Kelton
needed for his CUDA stuff and now I've adoped the same approach so we
can use one shared set of macros.

The netcdf files are now written in 2D Z slices in order to accommodate
the conventions we've adapted in the 3D arrays. A quick performance test
compaing to doing it in 1 3D array with and without compression turned
on showed negligible differences. Only a temporary 2D floating point
buffer was needed to do this memory-wise.