Now we read zfp_accuracy_LOFS from LOFS data, and write the values to the netcdf files, along with the chosen ZFP accuracy for the hdf2nc netcdf output file. To be clear: The point of all of this is to get the benefit from ZFP not just from LOFS but from the netCDF files that are created on the fly with hdf2nc. There is no reason to make the ZFP accuracy parameter smaller for the netCDF file than was set in CM1; in fact I think I'll make a check for that... and right now the ZFP accuracy parameter for the netcdf file is hardwired in the code. If you want to change these, you have to go and change the constant in set_nc_meta_zfp_name_units in our big if else ladder that loops through all the possible variables:
else if(same(var,"thrhopert")) set_nc_meta_zfp_name_units(5.0e-2,cmd->zfp,nid,v3did,"long_name","density_potential_temperature_perturbation","K");
                                                          ^^^^^^
This if/then/else ladder is where all of the variable-dependent stuff happens. It really
makes no sense to me to create another place where all of these are set, only to use a
uniquely named variable for the ZFP accuracy parameter.

There may be a more elegant way to do this in C but if there is I don't know what it is.
I'd really like a case statement but those only work with integers in C...

And to be abundantly clear, yes, we are compressing with CM1, uncompressing with hdf2nc,
and recompressing back to ZFP in the netCDF files. So long as our accuracy parameters
aren't stupid this should result in what we want w/out any artifats.

The moral of the story is to be aware of these compression values... and don't do anything
stupid! I send stuff to standard output/error to remind people of things (right now only
with --verbose). I'll probably fiddle a bit with this but the good news is, it works.