Lots of changs to the zfpwrite branch, some not related to ZFP but most are.
Latest discovery is that we need to set chunk dimensions to be evenly
divisible by 4 in all three dimensions and same for domain dimensions.
Otherwise netcdf decides what the chunk dimensions are and we get zeroes
around the boundary of our cube...
Will need to clean up and test before merging with master...

The good news is, we can write ZFP compressed data with hdf2nc pretty
cleanly.