Added recursive mkdir C function for making ncdir in case it needs to make nested directories. We need to spread our output amongst mulitple directories, much like LOFS, when writing. TACC threw a flag and stopped my hdf2nc batch jobs because of I/O issues with the machine.
Raised ZFP parameters. Need to have an external file for these. On the
list.

Added --nocmd option for not leaving .cmd files. Since I'm constructing
hdf2nc-commands with all the commands it's a bit redundant and adds exra
files. But the default state is to write those files - too convenient!