These changes are in response to the possibility (TX/TACC) that the cache files living on the Lustre share were causing performance issues (perhaps repeatedly reading these tiny files).
So I decided to have the option of writing the cache files to /dev/shm
instead of from the usual history file directory-relative spot. This is
meant for big high throughput runs where a lot of jobs are crunching.
This is likely just not important, but I kind of like the idea.

Since this means the 1st run with --devshmcache will always (well,
usually) create new cache files (because we're on some random compute node
that we've not run hdf2nc --devshmcache on yet) there will be an upfront
cost where processes create the cachefile. I've made the code a bit less
susceptible to race-condition creation of cache files. Also, I re-create
the entire directory structure, from slash (/), in /dev/shm/orf, i.e.:
/dev/shm/orf/scratch3/06618/orf/2021... blah blah blah
So as to ensure unique files taht don't conflict.

The second option I've added is --checkcmd, which will skip all
processing if a .cmd file is found for a .nc file. I am anticipating
having to re-run jobs because conversion is slow. This is especially
important because I'm now randomizing the hdf2nc-commands file for the
launcher process, because when everyone is hammering the same files
(when the times are sequential) this can also cause peformance issues.

However nothing will really fix the fact that Frontera's HT bandwidth
is... lacking.