OK I think I finally slayed this beast.
It really was all about not reading in that extra z level, both with the
sounding, but also with the readaheads

The problem was the thrhopert and perspert readaheads did not go to Z1+1
when reading in data even though we had the space allocated.

I was hunting in the wrong place for the error, but I think I finally
fixed this top layer crap. Will do tests with all of the interp
fields, some of which have been recently fucked for various different
reasons.