UNTESTED CODE - first check-in
New code that culls cm1hdf5 files. A long time ago I wrote something
similar. This code takes the --x0,--x1,--y0,--y1 mandatory arguments
that spans the data we want to KEEP, and writes out the commands to
delete the remaining files. So I can envision using ncview or similar to
bound the cube we want to keep, and passing those arguments to nukefiles
to write the script to delete the remainder. I might just have a system
command to delete the files in the code, but I will start with the
script out of paranoia. The purpose of this code is to deal with file
storage quotas on Frontera. Because I can't perfectly anticipate where
the cloud will be or what I want to keep before I run a simulation for
the first time, it's safest to save the full domain and cull later.
With most of my runs I've done a scouting run first to see if anything
interestig happens, and then re-run from restarts, saving only what I
need in the lofs namelist (hence, culling by not writing). But sometimes
I need to save data frequently up front (like with the anvil study)
because I don't have an infinite amount of computer time, and I know the
data that will be produce will be worth looking at (all supercells will
form anvils!).